requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-cors-headers==3.0.2
djangorestframework==3.9.4
idna==2.8
psycopg2==2.8.2
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
